+++ src/plugins/ELF_RSS2/parsing/routes/weibo.py
@@ -5,8 +5,9 @@ from pyquery import PyQuery as Pq
 
 from ...config import config
 from ...rss_class import Rss
-from .. import ParsingBase, get_summary, handle_html_tag
+from .. import ParsingBase, handle_html_tag
 from ..handle_images import handle_img_combo, handle_img_combo_with_content
+from ..utils import get_summary
 
 
 # 处理正文 处理网页 tag
